capsule component review
